refactor(test): move test to routers/public folder
